requirements.txt
aiosmtplib==3.0.2
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
fastapi-mail==1.5.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.49.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.6.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
